Migrate El Proxito views to class-based views
