+++ readthedocs/proxito/urls.py
@@ -33,25 +33,32 @@ from django.conf.urls import url
 from django.views import defaults
 
 from readthedocs.constants import pattern_opts
-from readthedocs.proxito import views
+from readthedocs.proxito.views.serve import (
+    ServeDocs,
+    ServeError404,
+    ServeRobotsTXT,
+    ServeSitemapXML,
+)
+from readthedocs.proxito.views.redirects import redirect_page_with_filename
+from readthedocs.proxito.views.utils import fast_404
 
 
 urlpatterns = [
     # Serve custom 404 pages
     url(
         r'^_proxito_404_(?P<proxito_path>.*)$',
-        views.serve_error_404,
+        ServeError404.as_view(),
         name='serve_error_404',
     ),
-    url(r'robots\.txt$', views.robots_txt, name='robots_txt'),
-    url(r'sitemap\.xml$', views.sitemap_xml, name='sitemap_xml'),
+    url(r'robots\.txt$', ServeRobotsTXT.as_view(), name='robots_txt'),
+    url(r'sitemap\.xml$', ServeSitemapXML.as_view(), name='sitemap_xml'),
 
     # # TODO: Support this?
     # (Sub)project `page` redirect
     url(
         r'^(?:projects/(?P<subproject_slug>{project_slug})/)?'
         r'page/(?P<filename>.*)$'.format(**pattern_opts),
-        views.redirect_page_with_filename,
+        redirect_page_with_filename,
         name='redirect_page_with_filename',
     ),
 
@@ -63,7 +70,7 @@ urlpatterns = [
             r'(?P<version_slug>{version_slug})/'
             r'(?P<filename>{filename_slug})$'.format(**pattern_opts)
         ),
-        views.serve_docs,
+        ServeDocs.as_view(),
         name='docs_detail',
     ),
 
@@ -85,11 +92,11 @@ urlpatterns = [
             r'^(?:projects/(?P<subproject_slug>{project_slug})/)?'
             r'(?P<filename>{filename_slug})$'.format(**pattern_opts)
         ),
-        views.serve_docs,
+        ServeDocs.as_view(),
         name='docs_detail_singleversion_subproject',
     ),
 ]
 
 # Use Django default error handlers to make things simpler
-handler404 = views.fast_404
+handler404 = fast_404
 handler500 = defaults.server_error
